Disable service manipulation via package hooks
By default, the "dh_installinit" packer helper will add logic to the
postinst, postrm, and prerm packaging hooks to manipulate the service
during package installtion, upgrade, and removal. As a result, the
service is automatically stopped during an upgrade, which is not what we
want to occur on a Delphix appliance. This change adds the required
options to "dh_installinit" to prevent this.